+++ carpetmodSrc/carpet/helpers/lifetime/spawning/JockeySpawningReason.java
@@ -1,15 +1,18 @@
 package carpet.helpers.lifetime.spawning;
 
+import carpet.helpers.lifetime.utils.JockeyType;
 import carpet.helpers.lifetime.utils.LifeTimeTrackerUtil;
 import carpet.utils.Messenger;
 import net.minecraft.entity.Entity;
 import net.minecraft.util.text.ITextComponent;
 
+import javax.annotation.Nonnull;
+
 public class JockeySpawningReason extends SpawningReason {
     private final Class<? extends Entity> partnerClass;
     private final JockeyType type;
 
-    public JockeySpawningReason(Entity partner, JockeyType type) {
+    public JockeySpawningReason(@Nonnull Entity partner, JockeyType type) {
         partnerClass = partner.getClass();
         this.type = type;
     }
@@ -33,17 +36,13 @@ public class JockeySpawningReason extends SpawningReason {
         switch (type) {
             case CHICKEN:
                 return Messenger.c("w Chicken jockey",
-                        "g  (with " + LifeTimeTrackerUtil.getEntityTypeDescriptor(partnerClass) + ")");
+                        "g  (" + LifeTimeTrackerUtil.getEntityTypeDescriptor(partnerClass) + ")");
             case SPIDER:
                 return Messenger.c("w Spider jockey");
             default:
                 return Messenger.c("w Unknown jockey",
-                        "g  (with " + LifeTimeTrackerUtil.getEntityTypeDescriptor(partnerClass) + ")");
+                        "g  (" + LifeTimeTrackerUtil.getEntityTypeDescriptor(partnerClass) + ")");
         }
     }
 
-    public enum JockeyType {
-        CHICKEN,
-        SPIDER
-    }
 }