+++ carpetmodSrc/carpet/helpers/lifetime/trackeddata/SizedTrackedData.java
@@ -1,4 +1,0 @@
-package carpet.helpers.lifetime.trackeddata;
-
-public class SizedTrackedData extends BasicTrackedData {
-}