+++ carpetmodSrc/carpet/helpers/lifetime/utils/JockeyType.java
@@ -1,0 +1,6 @@
+package carpet.helpers.lifetime.utils;
+
+public enum JockeyType {
+    CHICKEN,
+    SPIDER
+}